+++ src/components/friends/data-provider/test/provider/index.ts
@@ -6,7 +6,8 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-import Provider, { provider } from 'core/data';
+import type { MiddlewareParams } from 'core/request';
+import Provider, { DecodersMap, provider } from 'core/data';
 
 export * from 'core/data';
 
@@ -15,4 +16,14 @@ export default class FriendsDataProvider extends Provider {
 	static override request: typeof Provider.request = Provider.request({
 		cacheStrategy: 'never'
 	});
+
+	static override decoders: DecodersMap = {
+		get: [
+			(data: unknown, params: MiddlewareParams): Promise<unknown> => new Promise((resolve) => {
+				params.opts.meta.provider?.emitter.emit('friendsDataProviderDecoder');
+
+				setTimeout(() => resolve(data));
+			})
+		]
+	};
 }